Initial Checkin
This is a basic Netbiscuits example using geolocation to determine the
user's location and display Tweets nearby.  It also uses the Twitter
search API to display tweets having mobile in them.  This example
defaults to Tweets in the Atlanta area until the user clicks the link to
geolocate themselves.  Content is unfiltered and output from Twitter may
contain obscenities.